backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
